requirements.txt
# requirements.txt
Flask
gunicorn
line-bot-sdk